refactor: moved out of endpoints folder as it's not and endpoint
